# docs/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TreasuryServiceOrchestrator.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sub-accounts/{subAccountId}/linked-bank-accounts": {
      "post": {
        "tags": [
          "CreateLinkedBankAccount"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkedBankAccountRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkedBankAccountRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkedBankAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedBankAccountResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedBankAccountResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedBankAccountResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "LinkedBankAccounts"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkedBankAccountResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkedBankAccountResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LinkedBankAccountResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/sub-accounts/{subAccountId}/redemptions": {
      "post": {
        "tags": [
          "CreateRedemption"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedemptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedemptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedemptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Redemptions"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedemptionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedemptionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RedemptionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/sub-accounts/{subAccountId}/transfers": {
      "post": {
        "tags": [
          "CreateTransfer"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transfers"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/sub-accounts/{subAccountId}/deposit-addresses": {
      "get": {
        "tags": [
          "DepositAddresses"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositAddressResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositAddressResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositAddressResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DepositAddresses"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDepositAddressRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDepositAddressRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDepositAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DepositAddressResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositAddressResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositAddressResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sub-accounts/{subAccountId}/recipients": {
      "get": {
        "tags": [
          "Recipients"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipientResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RegisterRecipient"
        ],
        "parameters": [
          {
            "name": "subAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRecipientRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRecipientRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRecipientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateLinkedBankAccountRequest": {
        "required": [
          "beneficiaryName",
          "accountNumber",
          "routingNumber",
          "bankName",
          "billingName",
          "billingCity",
          "billingCountry",
          "billingLine1",
          "billingPostalCode",
          "billingLine2",
          "billingDistrict",
          "bankAddressCountry",
          "bankAddressBankName"
        ],
        "type": "object",
        "properties": {
          "beneficiaryName": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "routingNumber": {
            "type": "string"
          },
          "bankName": {
            "type": "string"
          },
          "billingName": {
            "type": "string"
          },
          "billingCity": {
            "type": "string"
          },
          "billingCountry": {
            "type": "string"
          },
          "billingLine1": {
            "type": "string"
          },
          "billingPostalCode": {
            "type": "string"
          },
          "billingLine2": {
            "type": [
              "null",
              "string"
            ]
          },
          "billingDistrict": {
            "type": [
              "null",
              "string"
            ]
          },
          "bankAddressCountry": {
            "type": "string"
          },
          "bankAddressBankName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateRedemptionRequest": {
        "required": [
          "linkedBankAccountId",
          "grossAmount"
        ],
        "type": "object",
        "properties": {
          "linkedBankAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "grossAmount": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
...
      "CreateTransferRequest": {
        "required": [
          "recipientId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "recipientId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "DepositAddressResponse": {
        "required": [
          "id",
          "subAccountId",
          "chain",
          "currency",
          "address",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "chain": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GenerateDepositAddressRequest": {
        "required": [
          "chain",
          "currency"
        ],
        "type": "object",
        "properties": {
          "chain": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "LinkedBankAccountResponse": {
        "required": [
          "id",
          "subAccountId",
          "beneficiaryName",
          "bankName",
          "circleBankAccountId",
          "status",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "beneficiaryName": {
            "type": "string"
          },
          "bankName": {
            "type": "string"
          },
          "circleBankAccountId": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/LinkedBankAccountStatus"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LinkedBankAccountStatus": {
        "type": "integer"
      },
      "Money": {
        "required": [
          "amount",
          "currencyCode"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currencyCode": {
            "type": "string"
          }
        }
      },
      "RecipientResponse": {
        "required": [
          "id",
          "subAccountId",
          "chain",
          "address",
          "label",
          "circleRecipientId",
          "status",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "chain": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "circleRecipientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/RecipientStatus"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RecipientStatus": {
        "type": "integer"
      },
      "RedemptionResponse": {
        "required": [
          "id",
          "subAccountId",
          "linkedBankAccountId",
          "grossAmount",
          "fees",
          "netAmount",
          "circleRedeemId",
          "status",
          "failureReason",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "linkedBankAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "grossAmount": {
            "$ref": "#/components/schemas/Money"
          },
          "fees": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Money"
              }
            ]
          },
          "netAmount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Money"
              }
            ]
          },
          "circleRedeemId": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TransferStatus"
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RegisterRecipientRequest": {
        "required": [
          "chain",
          "address",
          "label"
        ],
        "type": "object",
        "properties": {
          "chain": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
...
      "TransferResponse": {
        "required": [
          "id",
          "subAccountId",
          "recipientId",
          "amount",
          "circleTransferId",
          "status",
          "failureReason",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "recipientId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "circleTransferId": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TransferStatus"
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransferStatus": {
        "type": "integer"
      },
...
    }
  },
...
}